Improve the navbar logo story
